provided templates for moduling the program
