+++ HangmanGame/game_logic.cpp
@@ -1,0 +1,28 @@
+//
+//  game_logic.cpp
+//  HangmanGame
+//
+//  Created by Wladimir Pasko on 30.01.22.
+//
+
+#include "game_logic.hpp"
+
+#include <iostream>
+using namespace std;
+
+void displayGameDetails(int maxTries){
+    cout << "\n"
+    "        .-------------------------------------------------------------------------------.\n"
+    "        |      _      _                                                                  |\n"
+    "        |     | |    | |   __ _    _ __      __ _        /\\/\\      __ _    _ __          |\n"
+    "        |     | |----| |  / _  |  |  _ \\    / _` |      /    \\    / _  |  |  _  \\        |\n"
+    "        |     | |----| | | (_| |  | | | |  | (_| |     / /\\/\\ |  | (_| |  | | | |        |\n"
+    "        |     |_|    |_|  \\_ _ |  |_| |_|   \\__, |     \\/    \\/   \\_ _ |  |_| |_|        |\n"
+    "        |                                    |___/                                       |\n"
+    "        .--------------------------------------------------------------------------------.\n";
+
+    cout << "The purpose of this game is to guess an animal name, secretly chosen by the application\n\n";
+    cout << "You have to guess one letter at a time and you can have " << maxTries << " wrong attempts\n\n";
+    cout << "Enter a lower-case letter and don't forget to enter key after each guess\n\n";
+    cout << "Let's play the game!\n\n";
+}
