+++ HangmanGame/game_logic.hpp
@@ -1,0 +1,16 @@
+//
+//  game_logic.hpp
+//  HangmanGame
+//
+//  Created by Wladimir Pasko on 30.01.22.
+//
+
+#ifndef game_logic_hpp
+#define game_logic_hpp
+
+#include <stdio.h>
+using namespace std;
+
+void displayGameDetails(int);
+
+#endif /* game_logic_hpp */
